requirements.txt
pypdf==5.5.0
python-dotenv==1.0.0
openai==1.82.0
pandas==2.2.3
tinydb==4.8.2